Move navComponents into useNav composable. Update apiSpec. Make icon and url required. Add url sort for consistent ordering.
