+++ kolibri/plugins/learn/assets/src/my_downloads/views/MyDownloadsSideNavEntry.js
@@ -1,25 +1,19 @@
 import { UserKinds, NavComponentSections } from 'kolibri.coreVue.vuex.constants';
-import navComponents from 'kolibri.utils.navComponents';
+import registerNavItem from 'kolibri.utils.registerNavItem';
 import coreStrings from 'kolibri.utils.coreStrings';
 import urls from 'kolibri.urls';
 import plugin_data from 'plugin_data';
 
-const component = {
-  name: 'MyDownloadsSideNavEntry',
-  get url() {
-    return urls['kolibri:kolibri.plugins.learn:my_downloads']();
-  },
-  get label() {
-    return coreStrings.$tr('myDownloadsLabel');
-  },
-  icon: 'download',
-  role: UserKinds.LEARNER,
-  priority: 10,
-  section: NavComponentSections.ACCOUNT,
-};
-
 if (plugin_data.allowLearnerDownloads) {
-  navComponents.register(component);
+  registerNavItem({
+    get url() {
+      return urls['kolibri:kolibri.plugins.learn:my_downloads']();
+    },
+    get label() {
+      return coreStrings.$tr('myDownloadsLabel');
+    },
+    icon: 'download',
+    role: UserKinds.LEARNER,
+    section: NavComponentSections.ACCOUNT,
+  });
 }
-
-export default component;
