+++ kolibri/plugins/learn/assets/src/views/LearnSideNavEntry.js
@@ -1,11 +1,10 @@
-import navComponents from 'kolibri.utils.navComponents';
+import registerNavItem from 'kolibri.utils.registerNavItem';
 import urls from 'kolibri.urls';
 import coreStrings from 'kolibri.utils.coreStrings';
 import baseRoutes from '../routes/baseRoutes';
 import { learnStrings } from './commonLearnStrings';
 
-const sideNavConfig = {
-  name: 'LearnSideNavEntry',
+registerNavItem({
   get url() {
     return urls['kolibri:kolibri.plugins.learn:learn']();
   },
@@ -35,10 +34,5 @@ const sideNavConfig = {
     return learnStrings.$tr('learnLabel');
   },
   icon: 'learn',
-  priority: 10,
   bottomBar: true,
-};
-
-navComponents.register(sideNavConfig);
-
-export default sideNavConfig;
+});
