+++ kolibri/plugins/user_auth/assets/src/views/LoginSideNavEntry.js
@@ -1,10 +1,9 @@
 import { UserKinds, NavComponentSections } from 'kolibri.coreVue.vuex.constants';
-import navComponents from 'kolibri.utils.navComponents';
+import registerNavItem from 'kolibri.utils.registerNavItem';
 import urls from 'kolibri.urls';
 import coreStrings from 'kolibri.utils.coreStrings';
 
-const sideNavConfig = {
-  name: 'LoginSideNavEntry',
+registerNavItem({
   get url() {
     return urls['kolibri:kolibri.plugins.user_auth:user_auth']();
   },
@@ -13,10 +12,5 @@ const sideNavConfig = {
   },
   icon: 'login',
   role: UserKinds.ANONYMOUS,
-  priority: 10,
   section: NavComponentSections.ACCOUNT,
-};
-
-navComponents.register(sideNavConfig);
-
-export default sideNavConfig;
+});
