+++ kolibri/plugins/user_profile/assets/src/views/UserProfileSideNavEntry.js
@@ -1,10 +1,9 @@
 import { UserKinds, NavComponentSections } from 'kolibri.coreVue.vuex.constants';
-import navComponents from 'kolibri.utils.navComponents';
+import registerNavItem from 'kolibri.utils.registerNavItem';
 import urls from 'kolibri.urls';
 import coreStrings from 'kolibri.utils.coreStrings';
 
-const sideNavConfig = {
-  name: 'UserProfileSideNavEntry',
+registerNavItem({
   get url() {
     return urls['kolibri:kolibri.plugins.user_profile:user_profile']();
   },
@@ -13,10 +12,5 @@ const sideNavConfig = {
   },
   icon: 'person',
   role: UserKinds.LEARNER,
-  priority: 10,
   section: NavComponentSections.ACCOUNT,
-};
-
-navComponents.register(sideNavConfig);
-
-export default sideNavConfig;
+});
